// frontend/src/types/movie.ts
export type Movie = {
    id: number;
    title: string;
    score: number;
    count: number;
    image: string;
}

export type MoviePage = {
    content: Movie[];
    last: boolean;
    totalPages: number;
    totalElements: number;
    size: number;
    number: number;
    first: boolean;
    numberOfElements: number;
    empty: boolean;
}


// as estruturas de Movie e de MoviePage podem ser obtidas a partir do objeto retornado pela API (basta checar no Postman)